incomedistribution/requirements.txt
uWSGI==2.0.20
Django==3.2.4
requests==2.26.0
djangorestframework==3.13.1
django_uwsgi==0.2.2
numpy==1.19.5
pandas==1.1.5
openpyxl==3.0.9
SQLAlchemy==1.3.24




